Move container component to containers
